requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.36
pydantic==2.8.2
python-multipart==0.0.6
openai==1.3.7
reportlab==4.0.7
python-dateutil==2.8.2
jinja2==3.1.2
python-dotenv==1.0.0